Update confige script; add deps for https dns proxy with quic
Some deps should be installed when https-dns-proxy is installed [1].
So far, https-dns-proxy is not supporting quic protocol, but you can
always install package 'dnsproxy' [2] [3], that provides that.
This commit is just installing deps packages to enable quic protocol
on the main router when its available and improve the configure script.

[1] https://docs.openwrt.melmac.net/https-dns-proxy/#http3-quic-support
[2] https://forum.openwrt.org/t/dns-over-http-3/133092/24
[3] https://github.com/AdguardTeam/dnsproxy?tab=readme-ov-file#usage