Update bpi-r4-BE14-community.yaml
Suggested change to customise the naming scheme of the generated images to keep a better overview.

If you download a whole bunch of images for testing, it is easy to lose track. Clearly naming the files can bring back some clarity ;-)

This is just a suggested change only.
No bug or big functional improvement - accept if you think this is useful for you and/or community.
... or simply discard it if it offers no meaningful added value. ;-)

In addition, the selection of the branch from which the configuration is built via curl is further dynamised.